Added auth for event changes and created pipeline to remove existing files
