+++ backend/src/controllers/article.ts
@@ -53,17 +53,34 @@ export const updateArticle: RequestHandler = async (req, res, next) => {
     const { id } = req.params;
     validationErrorParser(errors);
 
-    // Returns the updated article
+    // Get the existing article to check for old thumbnail
+    const existingArticle = await ArticleModel.findById(id);
+    if (existingArticle === null) {
+      throw createHttpError(404, "Article not found at id " + id);
+    }
+
+    // Delete old thumbnail if it exists and is different from the new one
     const reqBody: ArticleUpdate = req.body as ArticleUpdate;
+    if (existingArticle.thumbnail && existingArticle.thumbnail !== reqBody.thumbnail) {
+      try {
+        const thumbnailUrl = existingArticle.thumbnail;
+        // Extract the path from the full URL
+        const path = thumbnailUrl.split("/o/")[1]?.split("?")[0];
+        if (path) {
+          const decodedPath = decodeURIComponent(path);
+          await firebaseStorage.bucket().file(decodedPath).delete();
+        }
+      } catch (error) {
+        console.error("Error deleting old thumbnail from Firebase:", error);
+        // Continue with article update even if thumbnail deletion fails
+      }
+    }
+
+    // Update the article
     const article = await ArticleModel.findByIdAndUpdate(id, reqBody, {
       new: true,
       runValidators: true,
     });
-
-    // Check if article of id exists
-    if (article === null) {
-      throw createHttpError(404, "Article not found at id " + id);
-    }
 
     res.status(200).json(article);
   } catch (error) {
